+++ src/adapters.zig
@@ -251,30 +251,41 @@ pub fn Peekable(comptime Wrapped: type) type {
         pub const Item = Wrapped.Item;
 
         iter: Iter(Wrapped),
-        peeked: ?Item = null,
+        peeked: ??Item = null,
         comptime _: markers.IsPeekable = .{},
 
         pub fn next(self: *Self) ?Item {
-            if (self.peeked) |_| {
-                const x = self.peeked;
-                self.peeked = null;
-                return x;
+            if (self.peeked) |peeked| {
+                if (peeked) |p| {
+                    self.peeked = null;
+                    return p;
+                }
             }
-
             return self.iter.next();
         }
 
         pub fn sizeHint(self: Self) struct { usize, ?usize } {
+            if (self.peeked) |peeked| {
+                if (peeked) |_| {
+                    var lower, var upper = self.iter.sizeHint();
+                    lower = math.add(usize, lower, 1) catch lower;
+                    if (upper) |u| upper = math.add(usize, u, 1) catch u;
+                    return .{ lower, upper };
+                }
+                return .{ 0, 0 };
+            }
             return self.iter.sizeHint();
         }
 
         pub fn peek(self: *Self) ?Item {
-            if (self.peeked) |opt_item| {
-                return opt_item;
+            if (self.peeked) |peeked| {
+                if (peeked) |p| {
+                    return p;
+                }
             }
 
             self.peeked = self.iter.next();
-            return self.peeked;
+            return self.peeked orelse null;
         }
     };
 }